requirements.txt
# python-telegram-bot==20.7
# pip install --upgrade python-telegram-bot
python-telegram-bot

python-dotenv==1.0.0
prettytable==3.9.0